I change pleace of AddressBook, I add Note and NotesBook + some functional for this classes
